allow to change image size via HTTP headers
If the request contains the headers X-Image-Width or
X-Image-Height, then the SVGG file's size will be adjusted.
Acceptable values for those headers are integers, e. g.:

    X-Image-Width: 900
    X-Image-Height: 500

If the headers do not contain a parseable integer, then they will
be ignored and fallback values will be used. Fallback for the
image width is 700 pixels, fallback for the image height is 400
pixels. This is the size that was used before this feature has
been introduced.