adjust Dockerfile for rare case when Debian stable has newer Node.js
In some rare cases, the Node.js version in Debian stable may be
more recent than the one provided by nodesource.com. In that
case, the version from the Debian stable repository will be
installed. But since Debian stable split the NPM binary into
another package, this will have to be installed separately.

This can be detected by attempting to run NPM, and if it fails,
then we are on the Debian stable repo's version and have to
install the Debian package for NPM, too.