fix error in handling of array data for graphs
Basically, plotly.js expects the data part of the input to be an
array. So the render function eagerly wrapped the data part in an
array. However, when the data part already was an array, it was
wrapped, too, resulting in an array containing an array. This
seemed to confuse plotly.js and resulted in wrongly rendered
plots.

That part's fixed now. The data is only wrapped in an array, if
it not is an array.